requirements.txt
# Core NLP
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Machine Learning
scikit-learn==1.4.0
transformers==4.38.2
torch==2.2.1
tensorflow==2.15.0

# Utilities
nltk==3.8.1
python-dotenv==1.0.0
gensim==4.3.2
numpy==1.26.0

# Gmail API
google-api-python-client==2.118.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0

# Your existing dev dependencies
pytest==7.4.4
black==23.12.1
flake8==6.1.0

#frontend
streamlit>=1.32.0
